+++ Assets/Prefabs/Slot.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 795b4aaafac822740a983661d0b0cc14
+timeCreated: 1425785853
+licenseType: Pro
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
