--- Assets/Scenes/Main.scene.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 69c8832e180d94113a569feb48b1510a
-timeCreated: 1425000722
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
