--- Assets/Scripts/GizmoUI.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 1aa1e5bacd0d9844ea291eab0100af22
-timeCreated: 1425242569
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
