+++ Assets/Scripts/Utils.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: e4d4d44de52045447a52effd5aead46a
+timeCreated: 1425781982
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
