+++ ProjectSettings/TagManager.asset
@@ -38,6 +38,9 @@ TagManager:
   - 
   - 
   m_SortingLayers:
+  - name: Background
+    uniqueID: 3033572447
+    locked: 0
   - name: Default
     uniqueID: 0
     locked: 0
